requirements.txt
Django==2.1.2
django-filter==2.0.0
djangorestframework==3.8.2
Markdown==3.0.1
Pygments==2.2.0
pytz==2018.5
